Plugin UI Sandbox for Linux
